Fixed Anthony Comments
Was forced to remove the login unit test for checking navigation to the home page occured. This is because for some reason the test never completes its subscription to the signIn event.
