+++ src/app/component/shared/footer/footer.component.html
@@ -8,7 +8,11 @@
         <p class="nav-link disabled m-0">Production v{{ version }}</p>
       </li>
       <li class="nav-item">
-        <a class="nav-link" routerLink="/statistics" routerLinkActive="active">
+        <a
+          class="nav-link"
+          routerLink="{{ routes.statistics.route }}"
+          routerLinkActive="active"
+        >
           Statistics
         </a>
       </li>
