+++ src/app/component/shared/footer/footer.component.ts
@@ -1,6 +1,7 @@
 import { ChangeDetectionStrategy, Component, OnInit } from "@angular/core";
 import { version } from "package.json";
 import { contactUsMenuItem } from "../../about/about.menus";
+import { statisticsMenuItem } from "../../statistics/statistics.menus";
 
 @Component({
   selector: "app-footer",
@@ -11,7 +12,8 @@ import { contactUsMenuItem } from "../../about/about.menus";
 export class FooterComponent implements OnInit {
   public version: string = version;
   routes = {
-    contactUs: contactUsMenuItem
+    contactUs: contactUsMenuItem,
+    statistics: statisticsMenuItem
   };
 
   constructor() {}
